+++ tests/e2e/ipWhitelist_spec.js
@@ -1,30 +1,30 @@
 const helpers = require("./global-setup");
 
-describe("ipWhitelist directive configuration", function () {
-	describe("Set ipWhitelist without access", function () {
-		beforeAll(function () {
+describe("ipWhitelist directive configuration", () => {
+	describe("Set ipWhitelist without access", () => {
+		beforeAll(() => {
 			helpers.startApplication("tests/configs/noIpWhiteList.js");
 		});
-		afterAll(async function () {
+		afterAll(async () => {
 			await helpers.stopApplication();
 		});
 
-		it("should return 403", function (done) {
+		it("should return 403", (done) => {
 			helpers.fetch(done, "http://localhost:8080").then((res) => {
 				expect(res.status).toBe(403);
 			});
 		});
 	});
 
-	describe("Set ipWhitelist []", function () {
-		beforeAll(function () {
+	describe("Set ipWhitelist []", () => {
+		beforeAll(() => {
 			helpers.startApplication("tests/configs/empty_ipWhiteList.js");
 		});
-		afterAll(async function () {
+		afterAll(async () => {
 			await helpers.stopApplication();
 		});
 
-		it("should return 200", function (done) {
+		it("should return 200", (done) => {
 			helpers.fetch(done, "http://localhost:8080").then((res) => {
 				expect(res.status).toBe(200);
 			});