+++ tests/e2e/modules/mocks/weather_forecast.js
@@ -4,7 +4,7 @@ const _ = require("lodash");
  * @param {object} extendedData extra data to add to the default mock data
  * @returns {string} mocked forecast weather data
  */
-function generateWeatherForecast(extendedData = {}) {
+const generateWeatherForecast = (extendedData = {}) => {
 	return JSON.stringify(
 		_.merge(
 			{},
@@ -110,6 +110,6 @@ function generateWeatherForecast(extendedData = {}) {
 			extendedData
 		)
 	);
-}
+};
 
 module.exports = generateWeatherForecast;