+++ CHANGELOG.md
@@ -15,7 +15,7 @@ _This release is scheduled to be released on 2023-01-01._
 
 ### Updated
 
-- updated e2e tests (moved `done()` in helper functions)
+- updated e2e tests (moved `done()` in helper functions) and use es6 syntax in all tests
 
 ### Fixed